+++ src/prettier.ts
@@ -2,7 +2,7 @@ import { Project } from './project';
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
 import { join } from 'path';
 import { ignore } from './ignore';
-import { npmInstall, npmRun } from './node-commands';
+import { npmInstall, npmRun, saveDevArgument } from './node-commands';
 import { writeError, writeIonic } from './logging';
 import { ExtensionContext, window } from 'vscode';
 import { Tip } from './tip';
@@ -17,11 +17,11 @@ export async function integratePrettier(project: Project) {
 
     if (question != 'Yes') return;
 
-    await project.run2(npmInstall('husky', '--save-dev', '--save-exact'));
+    await project.run2(npmInstall('husky', saveDevArgument(project), '--save-exact'));
     writeIonic(`Installed husky`);
-    await project.run2(npmInstall('prettier', '--save-dev', '--save-exact'));
+    await project.run2(npmInstall('prettier', saveDevArgument(project), '--save-exact'));
     writeIonic(`Installed prettier`);
-    await project.run2(npmInstall('lint-staged', '--save-dev', '--save-exact'));
+    await project.run2(npmInstall('lint-staged', saveDevArgument(project), '--save-exact'));
     writeIonic(`Installed lint-staged`);
     const filename = join(project.projectFolder(), 'package.json');
     const packageFile = JSON.parse(readFileSync(filename, 'utf8'));